HW02(base)-complete

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -1,15 +1,21 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
+@Repository
 public class InMemoryMealRepository implements MealRepository {
+    private static final Logger log = LoggerFactory.getLogger(InMemoryUserRepository.class);
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
@@ -17,7 +23,6 @@ public class InMemoryMealRepository implements MealRepository {
         MealsUtil.MEALS.forEach(this::save);
     }
 
-    @Override
     public Meal save(Meal meal) {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
@@ -29,18 +34,55 @@ public class InMemoryMealRepository implements MealRepository {
     }
 
     @Override
-    public boolean delete(int id) {
+    public Meal save(Meal meal, int userId) {
+        log.info("save {} from User {}", meal, userId);
+        if (meal.isNew()) {
+            meal.setId(counter.incrementAndGet());
+            repository.put(meal.getId(), meal);
+            return meal;
+        }
+        // handle case: update, but not present in storage
+        if (meal.getUserId().equals(userId)) {
+            return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        }
+        log.info("save {} from User {} - UserId mismatch", meal, userId);
+        return null;
+
+    }
+
+    @Override
+    public boolean delete(int id, int userId) {
+        Meal meal = get(id, userId);
+        if (meal == null) {
+            return false;
+        }
         return repository.remove(id) != null;
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int id, int userId) {
+        Meal meal = repository.get(id);
+        if (meal.getUserId().equals(userId)) {
+            return meal;
+        }
+        return null;
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public List<Meal> getAll(int userId) {
+        List<Meal> userMealList = repository.values().stream()
+                .filter(meal -> meal.getUserId().equals(userId)).collect(Collectors.toList());
+        userMealList.sort(new Comparator<Meal>() {
+            @Override
+            public int compare(Meal o1, Meal o2) {
+                if (o1.getDateTime().compareTo(o2.getDateTime()) > 0) {
+                    return -1;
+                } else {
+                    return 1;
+                }
+            }
+        });
+        return userMealList;
     }
 }
